Move into a different file
because I don't think Rod will like looking at it in the tests
